Reimported my buildings at a slightly larger scale to try and make them look bigger, maybe I just made this first iteration too small.
Added Crouching Functionality

Added a few more inputs that arent really used yet

Copied the raycast interact action for a grab action that allows picking up physics enabled objects and lugging them around with the player character. Made a interface for it as well but unfinished from here. figure we can get thet transform of the HeldObject and set it to the player characters pos plus an offset and just have it float around until you drop it or something like that maybe