shitnipples
Added new sound for anomaly that we can change again but figured it was a good placeholder in the meantime

went to fix sprinting bceause I thought it was broken but then I realized I guess it just doesnt work sometimes, still havent disected that issue. Seems to work most often

tried to add dps trigger to the anomaly but the cast kept failing, got late enough that felt I had to stope for the night, sorry for the broken bit. Im sure im just missing something small