poo
refined some things, arms shirt now has a transparency effect to make it more immersive. reweighted the ghoul skelemesh and modified model abit. tried fixing drone not sure if successfull. added fog back in not sure why it was gone, is suspect jeff or jim. added a name thing to main menu on new game which becomes the saveslot name.