Spafloofly
made it so hacking minigame resets when activated from seperate consoles

adjusted lockpicking logic a little to work with searchable/static interable bases. hooked it up to the safe in the grocery store

fixed bug where interaction UI would lag behind or get stuck when the hitcast would hit one interactable and then you mouseover straight to a different one

added filing cabinets to searchablebase and made the drawers move when searched

I think Ive found a bug with searchable base inventories but not sure yet, its definitely worked for me atleast once on an ammo container but the container window was bellow the vicinity window that time. any time its appeared above the vicinity I think its empty though or something maybe

realized some of our cardboard boxes all have interiors so exported one to blender and rigged it so we can have the lid open when we search it, added a searchablebase class for it

added icebox and dumpster to searchable base and made the lids rotate when searched

implemented basic burst fire mechanic from existing system